+++ src/examples/java/Main.java
@@ -13,23 +13,13 @@ import javax.security.auth.login.LoginException;
 // TODO: 04.08.2018 Write proper examples that make sense 
 public class Main {
 
-    private static CommandSettings settings;
-
     public static void main(String[] args) throws LoginException, InterruptedException {
         JDA jda = new JDABuilder(AccountType.BOT).setToken(Secrets.TOKEN).buildBlocking();
         // default prefix shall be "!" and we want the labels to be case insensitive.
-        settings = new CommandSettings("!", jda, true);
-        settings.setCooldown(3000) // commands can only be executed every 3 seconds now
-                //.addHelpLabels("help", "helpme") // help can now be demanded by calling one of these labels.
-                .put(new DefaultHelpCommand(), "help", "helpme")
-                .put(new PingCommand(), "ping")
-                .put(new PrefixCommand(), "setprefix")
-                .put(new KillCommand(settings), "kill", "deactivate", "shutdown") // setting aliases
+        new CommandSettings("!!", jda, true).setCooldown(3000) // commands can only be executed every 3 seconds now
+                .put(new PingCommand(), "ping", "p")
+                .put(new CustomPrefixCommand(), "prefix")
                 .activate(); // Activating! Very important!
-    }
-
-    public static String getPrefix(long guildId) {
-        return settings.getPrefix(guildId);
     }
 
 }