+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/AbstractCommand.java
@@ -10,9 +10,10 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 /**
- * This abstract class is an alternative implementation of ICommand.
+ * This abstract class isDefault an alternative implementation of ICommand.
  * Instead of implementing ICommand directly, you may create sub classes of this class and register them as commands.
  * This class provides the possibility to use sub commands, making it possible to annotate methods as SubCommand methods.
+ * To learn more about using sub commands, please refer to the readme on github.
  * @author JohnnyJayJay
  * @version 3.3
  * @see ICommand
@@ -27,9 +28,14 @@ public abstract class AbstractCommand implements ICommand {
         final Class<?>[] parameterTypes = {CommandEvent.class, Member.class, TextChannel.class, String[].class};
         this.subCommands = new HashMap<>();
         for (Method method : this.getClass().getDeclaredMethods()) {
-            if (method.isAnnotationPresent(SubCommand.class) && method.getReturnType().equals(Void.TYPE)
-                    && Modifier.isPublic(method.getModifiers()) && Arrays.equals(method.getParameterTypes(), parameterTypes)) {
-                subCommands.put(method.getAnnotation(SubCommand.class), method);
+            if (method.isAnnotationPresent(SubCommand.class)) {
+                if (method.getReturnType().equals(Void.TYPE) && Modifier.isPublic(method.getModifiers()) && Arrays.equals(method.getParameterTypes(), parameterTypes)) {
+                    subCommands.put(method.getAnnotation(SubCommand.class), method);
+                } else {
+                    CommandSettings.LOGGER.warn("You are using an invalid method signature for the SubCommand-annotation on method" + getClass().getName() + "#" + method.getName()
+                            + ".\nExpected: void (com.github.johnnyjayjay.commandapi.CommandEvent, net.dv8tion.jda.core.entities.Member, net.dv8tion.jda.core.entities.TextChannel, java.lang.String[])\nFound: "
+                            + method.getReturnType().getName() + " (" + Arrays.stream(method.getParameterTypes()).map(Class::getName).collect(Collectors.joining(", ")) + ")\nThis method will therefore be ignored.");
+                }
             }
         }
     }
@@ -38,23 +44,29 @@ public abstract class AbstractCommand implements ICommand {
     public final void onCommand(CommandEvent event, Member member, TextChannel channel, String[] args) {
         String joinedArgs = event.getCommandSettings().isLabelIgnoreCase() ? event.getCommand().getJoinedArgs().toLowerCase() : event.getCommand().getJoinedArgs();
         Member selfMember = event.getGuild().getSelfMember();
-        Set<SubCommand> possibleMethods = subCommands.keySet().stream().filter((sub) -> joinedArgs.startsWith(sub.name()))
-                .filter((sub) -> member.hasPermission(channel, sub.memberPerms()))
-                .filter((sub) -> selfMember.hasPermission(channel, sub.botPerms()))
-                .collect(Collectors.toSet());
-        bestMethod(possibleMethods, joinedArgs).ifPresent((method) -> {
-            try {
-                method.invoke(this, event, member, channel, args);
-            } catch (IllegalAccessException | InvocationTargetException e) {
-                CommandSettings.LOGGER.error("An Exception occurred while trying to invoke sub command method; Please report this in a github issue. https://github.com/JohnnyJayJay/discord-api-command/issues", e);
-            }
+        // FIXME: 06.08.2018 default method is always chosen
+        subCommands.keySet().stream()
+                .filter((sub) -> joinedArgs.matches(sub.regex() + ".*"))
+                .filter((sub) -> sub.argsLength() == args.length || sub.argsLength() == -1)
+                .filter((sub) -> event.checkMemberPermissions(sub.memberPerms()) && event.checkBotPermissions(sub.botPerms())).findFirst()
+                .map(Optional::of).orElseGet(
+                        () -> subCommands.keySet().stream().filter((sub) -> event.checkMemberPermissions(sub.memberPerms()) && event.checkBotPermissions(sub.botPerms())).filter(SubCommand::isDefault).findFirst())
+                .map(subCommands::get).ifPresent((method) -> {
+                    try {
+                        method.invoke(this, event, member, channel, args);
+                    } catch (IllegalAccessException e) {
+                        CommandSettings.LOGGER.error("An Exception occurred while trying to invoke sub command method; Please report this in a github issue. https://github.com/JohnnyJayJay/discord-api-command/issues", e);
+                    } catch (InvocationTargetException e) {
+                        CommandSettings.LOGGER.warn("One of the commands had an uncaught exception:", e.getCause());
+                        e.printStackTrace();
+                    }
         });
     }
 
-    private Optional<Method> bestMethod(Set<SubCommand> possible, String joinedArgs) {
-        Optional<Method> ret = Optional.empty();
+    /*private Optional<SubCommand> bestMethod(Set<SubCommand> possible, String joinedArgs) {
+        Optional<SubCommand> ret = Optional.empty();
         if (possible.size() == 1) {
-            ret = possible.stream().map(subCommands::get).findFirst();
+            ret = possible.stream().findFirst();
         } else if (!possible.isEmpty()) {
             SubCommand longest = null;
             int longestPrefixLength = 0;
@@ -87,11 +99,11 @@ public abstract class AbstractCommand implements ICommand {
                     }
                 }
             }
-            if (longest != null)
-                ret = Optional.of(subCommands.get(longest));
+            if (longest != null) {
+                ret = Optional.of(longest);
+            }
         }
         return ret;
-
-    }
+    }*/
 
 }