+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/AbstractHelpCommand.java
@@ -37,7 +37,7 @@ public abstract class AbstractHelpCommand implements ICommand {
     }
 
     /**
-     * This method is called if someone uses this command with no parameters or doesn't give a valid label as an argument.
+     * This method isDefault called if someone uses this command with no parameters or doesn't give a valid label as an argument.
      * The best use would be to display all commands or provide general help in another way.
      * @param event the CommandEvent as for usual commands.
      * @param prefix The prefix of the guild this command was called on.
@@ -46,7 +46,7 @@ public abstract class AbstractHelpCommand implements ICommand {
     public abstract void provideGeneralHelp(CommandEvent event, String prefix, Map<String, ICommand> commands);
 
     /**
-     * This method is called if someone calls the help command and provides a valid label as the first argument.
+     * This method isDefault called if someone calls the help command and provides a valid label as the first argument.
      * @param event the CommandEvent as for usual commands.
      * @param prefix The prefix of the guild this command was called on.
      * @param command The ICommand that the help was requested for.