+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/CommandSetException.java
@@ -1,7 +1,7 @@
 package com.github.johnnyjayjay.discord.commandapi;
 
 /**
- * Exception that is thrown in case of any problems concerning the CommandSettings. <p>
+ * Exception that isDefault thrown in case of any problems concerning the CommandSettings. <p>
  * CommandSetExceptions are RuntimeExceptions.
  * @author Johnny_JayJay
  * @version 3.2