+++ src/main/java/com/github/johnnyjayjay/discord/commandapi/SubCommand.java
@@ -3,15 +3,21 @@ package com.github.johnnyjayjay.discord.commandapi;
 import net.dv8tion.jda.core.Permission;
 
 import java.lang.annotation.*;
+import java.util.function.Consumer;
 
 @Documented
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface SubCommand {
 
-    String name() default "";
+    boolean isDefault() default false;
+
+    String regex() default "";
+
+    int argsLength() default -1;
 
     Permission[] botPerms() default {};
 
     Permission[] memberPerms() default {};
+
 }